Fixing issues from the review. FINALLY getting rid of some compiler warnings.

+++ code/IE/anchocommons/AnchoCommons/JSValueWrapper.hpp
@@ -394,7 +394,7 @@ public:
   }*/
 
   JSValueWrapperConst
-  operator[](int aIdx)const
+  operator[](size_t aIdx)const
   {
     return operator[](boost::lexical_cast<std::wstring>(aIdx));
   }
